restructuing adapter.py into smaller files
