+++ qiskit_braket_provider/providers/adapter.py
@@ -7,12 +7,9 @@ sequences that should not be optimized.
 """
 
 import warnings
-from collections import defaultdict
-from collections.abc import Callable, Collection, Iterable, Iterator, Mapping, Sequence
+from collections.abc import Callable, Collection, Iterable, Mapping, Sequence
 from dataclasses import dataclass
-from math import inf, pi, prod
-from numbers import Number
-from typing import Any, Self, TypeAlias, TypeVar, overload
+from typing import Any, TypeAlias, TypeVar, overload
 
 import numpy as np
 import qiskit.circuit.library as qiskit_gates
@@ -21,1198 +18,59 @@ from qiskit import QuantumCircuit, transpile
 from qiskit.circuit import (
     Barrier,
     BoxOp,
-    CircuitInstruction,
-    Clbit,
     ControlledGate,
-    ForLoopOp,
-    Gate,
-    IfElseOp,
     Measure,
     Parameter,
     ParameterExpression,
     ParameterVectorElement,
-    Qubit,
-    WhileLoopOp,
 )
 from qiskit.circuit import Instruction as QiskitInstruction
-from qiskit.circuit.library import get_standard_gate_name_mapping
-from qiskit.dagcircuit import DAGCircuit
 from qiskit.quantum_info import Pauli, SparsePauliOp
 from qiskit.transpiler import (
-    InstructionProperties,
     PassManager,
-    QubitProperties,
     Target,
-    TransformationPass,
 )
-from qiskit_ionq import add_equivalences, ionq_gates
-from sympy import Add, Expr, Mul, Pow, Symbol
+from qiskit_ionq import add_equivalences
 
-from braket import experimental_capabilities as braket_expcaps
-from braket.aws import AwsDevice, AwsDeviceType
+from braket.aws import AwsDevice
 from braket.circuits import Circuit, Instruction, measure
 from braket.circuits import Observable as BraketObservable
-from braket.circuits import gates as braket_gates
-from braket.circuits import noises as braket_noises
 from braket.circuits import observables as braket_observables
-from braket.default_simulator.openqasm._helpers.arrays import convert_range_def_to_range
-from braket.default_simulator.openqasm._helpers.casting import cast_to
-from braket.default_simulator.openqasm._helpers.functions import (
-    evaluate_binary_expression,
-    evaluate_unary_expression,
-)
-from braket.default_simulator.openqasm.interpreter import Interpreter, VerbatimBoxDelimiter
-from braket.default_simulator.openqasm.parser.openqasm_ast import (
-    ArrayLiteral,
-    BinaryExpression,
-    BinaryOperator,
-    BitType,
-    BooleanLiteral,
-    Cast,
-    ClassicalType,
-    DiscreteSet,
-    Expression,
-    FloatLiteral,
-    Identifier,
-    IndexedIdentifier,
-    IndexExpression,
-    IntegerLiteral,
-    RangeDefinition,
-    SymbolLiteral,
-    UnaryExpression,
-)
-from braket.default_simulator.openqasm.program_context import (
-    AbstractProgramContext,
-)
-from braket.device_schema import (
-    DeviceActionType,
-    DeviceCapabilities,
-    OpenQASMDeviceActionProperties,
-)
-from braket.device_schema.ionq import IonqDeviceCapabilities
-from braket.device_schema.rigetti import RigettiDeviceCapabilities, RigettiDeviceCapabilitiesV2
-from braket.device_schema.simulators import GateModelSimulatorDeviceCapabilities
-from braket.device_schema.standardized_gate_model_qpu_device_properties_v1 import (
-    StandardizedGateModelQpuDeviceProperties as StandardizedPropertiesV1,
-)
-from braket.device_schema.standardized_gate_model_qpu_device_properties_v2 import (
-    StandardizedGateModelQpuDeviceProperties as StandardizedPropertiesV2,
-)
-from braket.devices import Device, LocalSimulator
+from braket.default_simulator.openqasm.interpreter import Interpreter
+from braket.devices import Device
 from braket.ir.openqasm import Program
-from braket.ir.openqasm.modifiers import Control
 from braket.parametric import FreeParameter, FreeParameterExpression, Parameterizable
-from qiskit_braket_provider.exception import QiskitBraketException
-from qiskit_braket_provider.providers import braket_instructions
+from qiskit_braket_provider.providers.constants import (
+    _BRAKET_GATE_NAME_TO_QISKIT_GATE,
+    _BRAKET_SUPPORTED_NOISES,
+    _BRAKET_TO_QISKIT_NAMES,
+    _BRAKET_VERBATIM_BOX_NAME,
+    _CONTROLLED_GATES_BY_QUBIT_COUNT,  # noqa: F401
+    _EPS,
+    _PAULI_MAP,
+    _QISKIT_CONTROLLED_GATE_NAMES_TO_BRAKET_GATES,
+    _QISKIT_GATE_NAME_TO_BRAKET_GATE,
+    _TRANSPILER_GATE_SUBSTITUTES,  # noqa: F401
+)
+from qiskit_braket_provider.providers.qasm_context import (
+    QiskitProgramContext,
+    sympy_to_qiskit,
+)
+from qiskit_braket_provider.providers.target import (
+    SubstitutedTarget,
+    aws_device_to_target,
+    gateset_from_properties,  # noqa: F401
+    get_controlled_gateset,  # noqa: F401
+    local_simulator_to_target,
+    native_angle_restrictions,  # noqa: F401
+    native_gate_connectivity,  # noqa: F401
+    native_gate_set,  # noqa: F401
+)
 
 add_equivalences()
 
-_BRAKET_TO_QISKIT_NAMES = {
-    "u": "u",
-    "phaseshift": "p",
-    "cnot": "cx",
-    "x": "x",
-    "y": "y",
-    "z": "z",
-    "t": "t",
-    "ti": "tdg",
-    "s": "s",
-    "si": "sdg",
-    "v": "sx",
-    "vi": "sxdg",
-    "swap": "swap",
-    "iswap": "iswap",
-    "rx": "rx",
-    "ry": "ry",
-    "rz": "rz",
-    "xx": "rxx",
-    "yy": "ryy",
-    "zz": "rzz",
-    "i": "id",
-    "h": "h",
-    "cy": "cy",
-    "cz": "cz",
-    "ccnot": "ccx",
-    "cswap": "cswap",
-    "cphaseshift": "cp",
-    "ecr": "ecr",
-    "prx": "r",
-    "gpi": "gpi",
-    "gpi2": "gpi2",
-    "ms": "ms",
-    "gphase": "global_phase",
-    "unitary": "unitary",
-    "kraus": "kraus",
-}
-
-_CONTROLLED_GATES_BY_QUBIT_COUNT = {
-    1: {
-        "ch": "h",
-        "cs": "s",
-        "csdg": "sdg",
-        "csx": "sx",
-        "crx": "rx",
-        "cry": "ry",
-        "crz": "rz",
-        "ccz": "cz",
-    },
-    3: {"c3sx": "sx"},
-    inf: {"mcx": "cx"},
-}
-
-_ADDITIONAL_U_GATES = {"u1", "u2", "u3"}
-
-_EPS = 1e-10  # global variable used to chop very small numbers to zero
-
-_QISKIT_GATE_NAME_TO_BRAKET_GATE: dict[str, Callable] = {
-    "u1": lambda lam: [braket_gates.U(0, 0, lam)],
-    "u2": lambda phi, lam: [braket_gates.U(pi / 2, phi, lam)],
-    "u3": lambda theta, phi, lam: [braket_gates.U(theta, phi, lam)],
-    "u": lambda theta, phi, lam: [braket_gates.U(theta, phi, lam)],
-    "p": lambda angle: [braket_gates.PhaseShift(angle)],
-    "cp": lambda angle: [braket_gates.CPhaseShift(angle)],
-    "cx": lambda: [braket_gates.CNot()],
-    "x": lambda: [braket_gates.X()],
-    "y": lambda: [braket_gates.Y()],
-    "z": lambda: [braket_gates.Z()],
-    "t": lambda: [braket_gates.T()],
-    "tdg": lambda: [braket_gates.Ti()],
-    "s": lambda: [braket_gates.S()],
-    "sdg": lambda: [braket_gates.Si()],
-    "sx": lambda: [braket_gates.V()],
-    "sxdg": lambda: [braket_gates.Vi()],
-    "swap": lambda: [braket_gates.Swap()],
-    "rx": lambda angle: [braket_gates.Rx(angle)],
-    "ry": lambda angle: [braket_gates.Ry(angle)],
-    "rz": lambda angle: [braket_gates.Rz(angle)],
-    "rzz": lambda angle: [braket_gates.ZZ(angle)],
-    "id": lambda: [braket_gates.I()],
-    "h": lambda: [braket_gates.H()],
-    "cy": lambda: [braket_gates.CY()],
-    "cz": lambda: [braket_gates.CZ()],
-    "ccx": lambda: [braket_gates.CCNot()],
-    "cswap": lambda: [braket_gates.CSwap()],
-    "rxx": lambda angle: [braket_gates.XX(angle)],
-    "ryy": lambda angle: [braket_gates.YY(angle)],
-    "ecr": lambda: [braket_gates.ECR()],
-    "iswap": lambda: [braket_gates.ISwap()],
-    "r": lambda angle_1, angle_2: [braket_gates.PRx(angle_1, angle_2)],
-    # IonQ gates
-    "gpi": lambda turns: [braket_gates.GPi(2 * pi * turns)],
-    "gpi2": lambda turns: [braket_gates.GPi2(2 * pi * turns)],
-    "ms": lambda turns_1, turns_2, turns_3: [
-        braket_gates.MS(2 * pi * turns_1, 2 * pi * turns_2, 2 * pi * turns_3)
-    ],
-    "zz": lambda angle: [braket_gates.ZZ(2 * pi * angle)],
-    # Global phase
-    "global_phase": lambda phase: [braket_gates.GPhase(phase)],
-    "unitary": lambda operators: [braket_gates.Unitary(operators[0])],
-    "kraus": lambda operators: [braket_noises.Kraus(operators)],
-    "CCPRx": lambda angle_1, angle_2, feedback_key: [
-        braket_expcaps.iqm.classical_control.CCPRx(angle_1, angle_2, feedback_key)
-    ],
-    "MeasureFF": lambda feedback_key: [
-        braket_expcaps.iqm.classical_control.MeasureFF(feedback_key)
-    ],
-}
-
-_QISKIT_CONTROLLED_GATE_NAMES_TO_BRAKET_GATES: dict[str, Callable] = {
-    controlled_gate: _QISKIT_GATE_NAME_TO_BRAKET_GATE[base_gate]
-    for gate_map in _CONTROLLED_GATES_BY_QUBIT_COUNT.values()
-    for controlled_gate, base_gate in gate_map.items()
-}
-
-_STANDARD_GATE_NAME_MAPPING = get_standard_gate_name_mapping()
-
-_BRAKET_GATE_NAME_TO_QISKIT_GATE: dict[str, QiskitInstruction | None] = {
-    "u": qiskit_gates.UGate(Parameter("theta"), Parameter("phi"), Parameter("lam")),
-    "u1": qiskit_gates.U1Gate(Parameter("theta")),
-    "u2": qiskit_gates.U2Gate(Parameter("theta"), Parameter("lam")),
-    "u3": qiskit_gates.U3Gate(Parameter("theta"), Parameter("phi"), Parameter("lam")),
-    "h": qiskit_gates.HGate(),
-    "ccnot": qiskit_gates.CCXGate(),
-    "cnot": qiskit_gates.CXGate(),
-    "cphaseshift": qiskit_gates.CPhaseGate(Parameter("theta")),
-    "cswap": qiskit_gates.CSwapGate(),
-    "cy": qiskit_gates.CYGate(),
-    "cz": qiskit_gates.CZGate(),
-    "i": qiskit_gates.IGate(),
-    "phaseshift": qiskit_gates.PhaseGate(Parameter("theta")),
-    "rx": qiskit_gates.RXGate(Parameter("theta")),
-    "ry": qiskit_gates.RYGate(Parameter("theta")),
-    "rz": qiskit_gates.RZGate(Parameter("phi")),
-    "s": qiskit_gates.SGate(),
-    "si": qiskit_gates.SdgGate(),
-    "swap": qiskit_gates.SwapGate(),
-    "t": qiskit_gates.TGate(),
-    "ti": qiskit_gates.TdgGate(),
-    "v": qiskit_gates.SXGate(),
-    "vi": qiskit_gates.SXdgGate(),
-    "x": qiskit_gates.XGate(),
-    "xx": qiskit_gates.RXXGate(Parameter("theta")),
-    "y": qiskit_gates.YGate(),
-    "yy": qiskit_gates.RYYGate(Parameter("theta")),
-    "z": qiskit_gates.ZGate(),
-    "zz": qiskit_gates.RZZGate(Parameter("theta")),
-    "ecr": qiskit_gates.ECRGate(),
-    "prx": qiskit_gates.RGate(Parameter("theta"), Parameter("phi")),
-    "iswap": qiskit_gates.iSwapGate(),
-    "gpi": ionq_gates.GPIGate(Parameter("phi") / (2 * pi)),
-    "gpi2": ionq_gates.GPI2Gate(Parameter("phi") / (2 * pi)),
-    "ms": ionq_gates.MSGate(
-        Parameter("phi0") / (2 * pi),
-        Parameter("phi1") / (2 * pi),
-        Parameter("theta") / (2 * pi),
-    ),
-    "gphase": qiskit_gates.GlobalPhaseGate(Parameter("theta")),
-    "measure": qiskit_gates.Measure(),
-    "unitary": qiskit_gates.UnitaryGate,
-    "kraus": qiskit_qi.Kraus,
-    "cc_prx": braket_instructions.CCPRx(
-        Parameter("angle_1"), Parameter("angle_2"), Parameter("feedback_key")
-    ),
-    "measure_ff": braket_instructions.MeasureFF(Parameter("feedback_key")),
-}
-
-_BRAKET_SUPPORTED_NOISES = [
-    "kraus",
-    "bitflip",
-    "depolarizing",
-    "amplitudedamping",
-    "generalizedamplitudedamping",
-    "phasedamping",
-    "phaseflip",
-    "paulichannel",
-    "twoqubitdepolarizing",
-    "twoqubitdephasing",
-    # "twoqubitpaulichannel" no to_openqasm support yet
-]
-
-_TRANSPILER_GATE_SUBSTITUTES: dict[tuple[str, tuple[float | str, ...]], Gate] = {
-    ("rx", (pi,)): qiskit_gates.XGate(),
-    ("rx", (-pi,)): qiskit_gates.XGate(),
-    ("rx", (pi / 2,)): qiskit_gates.SXGate(),
-    ("rx", (-pi / 2,)): qiskit_gates.SXdgGate(),
-}
-
-_PAULI_MAP = {
-    "X": braket_observables.X,
-    "Y": braket_observables.Y,
-    "Z": braket_observables.Z,
-}
-
-_BRAKET_VERBATIM_BOX_NAME = "verbatim"
-
 _Translatable: TypeAlias = QuantumCircuit | Circuit | Program | str
-_ParamKey: TypeAlias = tuple[float | str, ...]
-_QubitSet: TypeAlias = set[tuple[int, ...]]
-_ParameterRestrictions: TypeAlias = dict[str, dict[_ParamKey, _QubitSet]]
-
 _T = TypeVar("_T")
-
-
-class _SubstitutedTarget(Target):
-    def __new__(cls, *args, **kwargs) -> Self:
-        out = super().__new__(cls, *args, **kwargs)
-        gate_substitutes: dict[str, dict[tuple[int, ...], QiskitInstruction]] = {}
-        out._gate_substitutes = gate_substitutes
-        out._pass_manager = PassManager([_SubstituteGates(gate_substitutes)])
-        return out
-
-    def _substitute(
-        self, circuits: QuantumCircuit | Iterable[QuantumCircuit]
-    ) -> QuantumCircuit | list[QuantumCircuit]:
-        single = isinstance(circuits, QuantumCircuit)
-        circuit_list: list[QuantumCircuit] = [circuits] if single else list(circuits)
-        results: list[QuantumCircuit] = self._pass_manager.run(circuit_list)
-        for original, result in zip(circuit_list, results, strict=True):
-            if original._layout is not None:
-                result._layout = original._layout
-        return results[0] if single else results
-
-
-class _SubstituteGates(TransformationPass):
-    def __init__(
-        self, gate_substitutes: Mapping[str, Mapping[tuple[int, ...], QiskitInstruction]]
-    ) -> None:
-        super().__init__()
-        self._gate_substitutes = gate_substitutes
-
-    def run(self, dag: DAGCircuit) -> DAGCircuit:
-        if not self._gate_substitutes:
-            return dag
-        qubits = {q: i for i, q in enumerate(dag.qubits)}
-        for node in dag.op_nodes():
-            if (op_name := node.op.name) in self._gate_substitutes:
-                dag.substitute_node(
-                    node, self._gate_substitutes[op_name][tuple(qubits[q] for q in node.qargs)]
-                )
-        return dag
-
-
-class _QiskitProgramContext(AbstractProgramContext):
-    """Program context for converting OpenQASM 3 programs to Qiskit circuits.
-
-    This context extends AbstractProgramContext to build Qiskit QuantumCircuits from
-    OpenQASM 3 programs. It supports Braket verbatim pragmas, which are converted to
-    Qiskit BoxOp operations to preserve gate sequences that should not be optimized.
-
-    Verbatim boxes are represented using Qiskit's native BoxOp construct, which treats
-    a block of operations atomically. All verbatim boxes in a circuit use the same
-    configurable label name.
-    """
-
-    num_qubits: int
-
-    def __init__(self, verbatim_box_name: str = _BRAKET_VERBATIM_BOX_NAME) -> None:
-        """Initialize the Qiskit program context.
-
-        Args:
-            verbatim_box_name: Name to use for BoxOp labels when converting verbatim boxes.
-                Default: "verbatim"
-        """
-        super().__init__()
-        self._circuit_stack: list[QuantumCircuit] = [QuantumCircuit()]
-        self._param_map: dict[str, Parameter] = {}
-        self._in_verbatim_box = False
-        self._verbatim_circuit: QuantumCircuit | None = None
-        self._verbatim_box_name = verbatim_box_name
-        self._clbit_offset: dict[str, int] = {}
-
-    @property
-    def _active_circuit(self) -> QuantumCircuit:
-        """The circuit that instructions should be added to (top of stack)."""
-        return self._circuit_stack[-1]
-
-    @property
-    def circuit(self) -> QuantumCircuit:
-        if self._in_verbatim_box:
-            raise ValueError(
-                "Unclosed verbatim box at end of program. "
-                "Every verbatim box start marker must have a matching end marker."
-            )
-        return self._circuit_stack[0]
-
-    def add_qubits(self, name: str, num_qubits: int | None = 1) -> None:
-        super().add_qubits(name, num_qubits)
-        self._active_circuit.add_register(num_qubits)
-
-    def declare_variable(
-        self,
-        name: str,
-        symbol_type: ClassicalType | type,
-        value: Any = None,  # noqa: ANN401
-        const: bool = False,
-    ) -> None:
-        """Override to add classical bits to the Qiskit circuit when declared.
-
-        When a classical bit array is declared (e.g., bit[2] c;), we need to add
-        the corresponding classical bits to the Qiskit circuit.
-
-        Note: This only adds classical bits when the size is known at declaration time.
-        For function parameters with variable sizes (e.g., bit[n] where n is a parameter),
-        the classical bits are not added since the size is not yet determined.
-        """
-        super().declare_variable(name, symbol_type, value, const)
-
-        # If this is a bit type declaration, add classical bits to the circuit
-        if isinstance(symbol_type, BitType):
-            if symbol_type.size is not None:
-                if isinstance(symbol_type.size, IntegerLiteral):
-                    size = symbol_type.size.value
-                else:
-                    # Size is an Identifier or expression, can't determine size yet
-                    # This happens for function parameters like bit[n] where n is a variable
-                    return
-            else:
-                size = 1
-
-            # this is used to deal with Qiskit's QuantumCircuit storing all classical bits in a flat list
-            self._clbit_offset[name] = self._active_circuit.num_clbits
-            self._active_circuit.add_bits([Clbit() for _ in range(size)])
-
-    def is_builtin_gate(self, name: str) -> bool:
-        return name in _BRAKET_GATE_NAME_TO_QISKIT_GATE
-
-    def add_phase_instruction(self, target: int | list[int], phase_value: float) -> None:  # noqa: ARG002
-        self._active_circuit.global_phase += phase_value
-
-    def add_gate_instruction(
-        self,
-        gate_name: str,
-        target: tuple[int, ...],
-        params: list[Any] | None,
-        ctrl_modifiers: list[int],
-        power: int,
-    ) -> None:
-        gate: Gate = _BRAKET_GATE_NAME_TO_QISKIT_GATE[gate_name].copy()
-        params = (
-            [float(param) if isinstance(param, Number) else param for param in params]  # type: ignore[arg-type]
-            if params is not None
-            else []
-        )
-        if params:
-            gate.params = params
-        gate = gate.power(float(power)) if power != 1 else gate
-        if ctrl_modifiers:
-            gate = gate.control(
-                len(ctrl_modifiers), ctrl_state=str("".join([str(i) for i in ctrl_modifiers]))
-            )
-
-        active = self._active_circuit
-        # Ensure circuit has enough qubits for the target indices by adding missing qubits
-        # This is needed when using physical qubits ($0, $1, etc.) where no qubit register is declared
-        max_qubits = (max(target) + 1) if target else -1
-        num_missing_qubits = max_qubits - active.num_qubits
-        active.add_bits([Qubit() for _ in range(num_missing_qubits)])
-        self.num_qubits = max(self.num_qubits, active.num_qubits)
-
-        if self._in_verbatim_box:
-            # Ensure verbatim circuit also has enough qubits by adding missing qubits
-            num_missing_qubits = max_qubits - self._verbatim_circuit.num_qubits
-            self._verbatim_circuit.add_bits([Qubit() for _ in range(num_missing_qubits)])
-            self._verbatim_circuit.append(CircuitInstruction(gate, target))
-        else:
-            active.append(CircuitInstruction(gate, target))
-
-    def handle_parameter_value(self, value: Number | Expr) -> Number | Parameter:
-        return _sympy_to_qiskit(value, self._param_map) if isinstance(value, Expr) else value
-
-    def add_measure(
-        self,
-        target: tuple[int],
-        classical_targets: Sequence[int] | None = None,
-        *,
-        classical_destination: Identifier | IndexedIdentifier | None = None,  # noqa: ARG002
-    ) -> None:
-        active = self._active_circuit
-        # this is to cover the edge case where a user measures a qubit without assigning it to a classical register
-        if active.num_clbits < len(target):
-            num_missing_clbits = len(target) - active.num_clbits
-            active.add_bits([Clbit() for _ in range(num_missing_clbits)])
-        for idx, qubit in enumerate(target):
-            index = classical_targets[idx] if classical_targets else idx
-            active.measure(qubit, index)
-
-    def add_verbatim_marker(self, marker: VerbatimBoxDelimiter) -> None:
-        """Handle verbatim box start/end markers.
-
-        When START_VERBATIM is received:
-        - Create a new QuantumCircuit to collect verbatim gates
-        - Set _in_verbatim_box flag to True
-
-        When END_VERBATIM is received:
-        - Wrap the collected gates in a BoxOp
-        - Append the BoxOp to the main circuit
-        - Reset verbatim state
-
-        Args:
-            marker: VerbatimBoxDelimiter indicating START_VERBATIM or END_VERBATIM
-
-        Raises:
-            ValueError: If nested verbatim boxes are encountered or if END_VERBATIM is called without START_VERBATIM
-        """
-
-        if marker == VerbatimBoxDelimiter.START_VERBATIM:
-            if self._in_verbatim_box:
-                raise ValueError("Nested verbatim boxes are not supported")
-
-            self._verbatim_circuit = QuantumCircuit()
-            self._in_verbatim_box = True
-
-        elif marker == VerbatimBoxDelimiter.END_VERBATIM:
-            if not self._in_verbatim_box:
-                raise ValueError("Verbatim box end marker without matching start")
-
-            box_op = BoxOp(self._verbatim_circuit, label=self._verbatim_box_name)
-
-            active = self._active_circuit
-            # Append BoxOp to active circuit with all qubits (convert indices to Qubit objects)
-            qubit_objects = [active.qubits[i] for i in range(self._verbatim_circuit.num_qubits)]
-            active.append(box_op, qubit_objects)
-
-            self._in_verbatim_box = False
-            self._verbatim_circuit = None
-
-        else:
-            raise ValueError("Verbatim box created using invalid marker")
-
-    @property
-    def supports_midcircuit_measurement(self) -> bool:
-        return True
-
-    def is_mcm_dependent(self, expression: Expression) -> bool:
-        """Check if expression depends on a mid-circuit measurement result.
-
-        Delegates to the base class for identifier-based checks using
-        _mcm_dependent_scopes, but always returns True for RangeDefinition
-        and DiscreteSet to force for-loops through evaluate_for_range
-        (producing ForLoopOp).
-        """
-        if isinstance(expression, (RangeDefinition, DiscreteSet)):
-            return True
-        return super().is_mcm_dependent(expression)
-
-    def iter_classical_scopes(self, expression: Expression):  # noqa: ARG002
-        """Yield once since Qiskit circuit building doesn't do per-path branching."""
-        yield
-
-    def evaluate_condition(self, condition: Expression) -> Iterator[bool]:
-        """Evaluate a branching condition using a circuit stack.
-
-        Yields True (visit if-block) then False (visit else-block).
-        Each yield pushes a new circuit onto the stack; after the interpreter
-        visits the block, the circuit is popped and used to build an IfElseOp.
-        """
-        # MCM path: resolve condition to (Clbit, value)
-        main = self._active_circuit
-        if isinstance(condition, (Identifier, IndexExpression)):
-            # Bare identifier like `if (c)` or `if (c[0])` — equivalent to `== 1`
-            resolved_condition = self._resolve_condition_from_identifier(condition)
-        elif isinstance(condition, BinaryExpression):
-            if condition.op != BinaryOperator["=="]:
-                raise NotImplementedError(
-                    f"Unsupported operator '{condition.op.name}' in branching condition. "
-                    f"Only '==' is supported for mid-circuit measurement branching."
-                )
-            resolved_condition = self._resolve_condition(condition)
-        else:
-            raise NotImplementedError(
-                f"Unsupported condition type '{type(condition).__name__}' in branching condition. "
-                f"Only Identifier, IndexExpression, and BinaryExpression (==) are supported."
-            )
-
-        true_body = QuantumCircuit(main.num_qubits, main.num_clbits)
-        self._circuit_stack.append(true_body)
-        yield True
-        self._circuit_stack.pop()
-
-        # Push circuit for else-block (the interpreter always consumes this yield
-        # even for if-only blocks; the empty circuit is discarded below)
-        false_body = QuantumCircuit(main.num_qubits, main.num_clbits)
-        self._circuit_stack.append(false_body)
-        yield False
-        self._circuit_stack.pop()
-
-        actual_false = false_body if false_body.data else None
-
-        if not true_body.data and not actual_false:
-            raise ValueError(
-                "Branching statement conditioned on a measurement has empty bodies. "
-                "Both if and else branches contain no quantum operations."
-            )
-
-        # Sync main circuit dimensions if branch bodies grew
-        max_qubits = max(true_body.num_qubits, false_body.num_qubits)
-        max_clbits = max(true_body.num_clbits, false_body.num_clbits)
-        if max_qubits > main.num_qubits:
-            main.add_bits([Qubit() for _ in range(max_qubits - main.num_qubits)])
-        if max_clbits > main.num_clbits:
-            main.add_bits([Clbit() for _ in range(max_clbits - main.num_clbits)])
-
-        if_else_op = IfElseOp(resolved_condition, true_body, actual_false)
-        qubits = list(range(max_qubits))
-        clbits = list(range(max_clbits))
-        main.append(if_else_op, qubits, clbits)
-
-    def evaluate_for_range(
-        self, set_declaration: Expression, loop_var: str, loop_type: ClassicalType
-    ) -> Iterator[None]:
-        """Capture the for-loop body into a ForLoopOp.
-
-        Yields once to capture the body. If the body contains quantum operations,
-        wraps it in a ForLoopOp. If purely classical (empty body circuit),
-        falls back to static unrolling for remaining iterations.
-        """
-        index = self._evaluate_expression(set_declaration)
-        if isinstance(index, RangeDefinition):
-            index_values = [IntegerLiteral(x) for x in convert_range_def_to_range(index)]
-        else:
-            index_values = index.values
-
-        if not index_values:
-            return
-
-        main = self._active_circuit
-        # First pass: capture with concrete value to detect classical vs quantum body
-        # Snapshot outer scope variables to detect classical side effects
-        outer_vars = dict(self.variable_table.current_scope)
-        probe = QuantumCircuit(main.num_qubits, main.num_clbits)
-        self._circuit_stack.append(probe)
-        with self.enter_scope():
-            self.declare_variable(loop_var, loop_type, index_values[0])
-            yield
-        self._circuit_stack.pop()
-
-        if not probe.data:
-            # Purely classical loop body — statically unroll remaining iterations
-            for i in index_values[1:]:
-                with self.enter_scope():
-                    self.declare_variable(loop_var, loop_type, i)
-                    yield
-            return
-
-        # Detect classical side effects: if outer scope variables changed during
-        # the probe, the body mutates classical state that ForLoopOp won't capture
-        modified_vars = [
-            name
-            for name, val in self.variable_table.current_scope.items()
-            if name in outer_vars and outer_vars[name] != val
-        ]
-        if modified_vars:
-            raise ValueError(
-                f"For loop body modifies classical variable(s) {modified_vars} "
-                f"which cannot be captured by ForLoopOp. "
-                f"Only quantum operations are preserved across loop iterations."
-            )
-
-        # Second pass: re-capture with symbolic loop variable for correct ForLoopOp binding
-        body = QuantumCircuit(main.num_qubits, main.num_clbits)
-        self._circuit_stack.append(body)
-        with self.enter_scope():
-            self.declare_variable(loop_var, loop_type, SymbolLiteral(Symbol(loop_var)))
-            yield
-        self._circuit_stack.pop()
-
-        indexset = tuple(iv.value for iv in index_values)
-        loop_param = self._param_map.get(loop_var) or Parameter(loop_var)
-        for_op = ForLoopOp(indexset, loop_param, body)
-        max_qubits = max(body.num_qubits, main.num_qubits)
-        max_clbits = max(body.num_clbits, main.num_clbits)
-        if max_qubits > main.num_qubits:
-            main.add_bits([Qubit() for _ in range(max_qubits - main.num_qubits)])
-        if max_clbits > main.num_clbits:
-            main.add_bits([Clbit() for _ in range(max_clbits - main.num_clbits)])
-        main.append(for_op, list(range(max_qubits)), list(range(max_clbits)))
-
-    def handle_loop_break(self):
-        """Reject break statements since Qiskit's ForLoopOp/WhileLoopOp do not support them."""
-        raise NotImplementedError("break statements are not supported in loops.")
-
-    def handle_loop_continue(self):
-        """Reject continue statements since Qiskit's ForLoopOp/WhileLoopOp do not support them."""
-        raise NotImplementedError("continue statements are not supported in loops.")
-
-    def evaluate_while_condition(self, condition: Expression) -> Iterator[bool]:
-        """Evaluate a while-loop condition, capturing the body into a WhileLoopOp."""
-        # MCM path: resolve condition and capture body into WhileLoopOp
-        main = self._active_circuit
-        if isinstance(condition, (Identifier, IndexExpression)):
-            resolved_condition = self._resolve_condition_from_identifier(condition)
-        elif isinstance(condition, BinaryExpression):
-            if condition.op != BinaryOperator["=="]:
-                raise NotImplementedError(
-                    f"Unsupported operator '{condition.op.name}' in while-loop condition. "
-                    f"Only '==' is supported for mid-circuit measurement while loops."
-                )
-            resolved_condition = self._resolve_condition(condition)
-        else:
-            raise NotImplementedError(
-                f"Unsupported condition type '{type(condition).__name__}' in while-loop condition. "
-                f"Only Identifier, IndexExpression, and BinaryExpression (==) are supported."
-            )
-
-        body = QuantumCircuit(main.num_qubits, main.num_clbits)
-        self._circuit_stack.append(body)
-        yield True
-        self._circuit_stack.pop()
-
-        if not body.data:
-            raise ValueError(
-                "While loop conditioned on a measurement has an empty body. "
-                "This would result in an infinite loop."
-            )
-
-        max_qubits = max(body.num_qubits, main.num_qubits)
-        max_clbits = max(body.num_clbits, main.num_clbits)
-        if max_qubits > main.num_qubits:
-            main.add_bits([Qubit() for _ in range(max_qubits - main.num_qubits)])
-        if max_clbits > main.num_clbits:
-            main.add_bits([Clbit() for _ in range(max_clbits - main.num_clbits)])
-
-        while_op = WhileLoopOp(resolved_condition, body)
-        main.append(while_op, list(range(max_qubits)), list(range(max_clbits)))
-
-    def _evaluate_expression(self, expression: Expression | list[Expression]) -> Any:  # noqa: ANN401
-        """Lightweight expression evaluator for loop conditions and ranges."""
-        match expression:
-            case (
-                BooleanLiteral()
-                | IntegerLiteral()
-                | FloatLiteral()
-                | ArrayLiteral()
-                | SymbolLiteral()
-            ):
-                return expression
-            case Identifier():
-                return self.get_value_by_identifier(expression)
-            case BinaryExpression(lhs=lhs, rhs=rhs, op=op):
-                return evaluate_binary_expression(
-                    self._evaluate_expression(lhs),
-                    self._evaluate_expression(rhs),
-                    op,
-                )
-            case UnaryExpression(expression=inner, op=op):
-                return evaluate_unary_expression(self._evaluate_expression(inner), op)
-            case Cast(type=cast_type, argument=argument):
-                return cast_to(cast_type, self._evaluate_expression(argument))
-            case RangeDefinition(start=start, end=end, step=step):
-                return RangeDefinition(
-                    self._evaluate_expression(start) if start else None,
-                    self._evaluate_expression(end),
-                    self._evaluate_expression(step) if step else None,
-                )
-            case DiscreteSet(values=values):
-                return DiscreteSet(values=[self._evaluate_expression(v) for v in values])
-            case list():
-                return [self._evaluate_expression(item) for item in expression]
-            case _:
-                raise TypeError(f"Cannot evaluate expression of type {type(expression).__name__}")
-
-    def _resolve_condition(self, condition: BinaryExpression) -> tuple[Clbit, int]:
-        """Convert an OpenQASM condition AST node to a Qiskit (Clbit, int) condition."""
-        if isinstance(condition.lhs, (Identifier, IndexExpression)):
-            clbit_index = self._resolve_clbit_index(condition.lhs)
-            value = condition.rhs.value
-        else:
-            clbit_index = self._resolve_clbit_index(condition.rhs)
-            value = condition.lhs.value
-        return (self.circuit.clbits[clbit_index], int(value))
-
-    def _resolve_condition_from_identifier(
-        self, condition: Identifier | IndexExpression
-    ) -> tuple[Clbit, int]:
-        """Convert a bare identifier condition (e.g., `c` or `c[0]`) to (Clbit, 1)."""
-        clbit_index = self._resolve_clbit_index(condition)
-        return (self.circuit.clbits[clbit_index], 1)
-
-    def _resolve_clbit_index(self, node: Identifier | IndexExpression) -> int:
-        """Resolve an identifier or indexed identifier to a classical bit index."""
-        if isinstance(node, IndexExpression):
-            name = node.collection.name
-            index = node.index[0].value
-        elif isinstance(node, Identifier):
-            name = node.name
-            var_type = self.get_type(name)
-            if isinstance(var_type, BitType) and var_type.size is not None:
-                size = var_type.size.value if isinstance(var_type.size, IntegerLiteral) else None
-                if size is not None and size > 1:
-                    raise TypeError(
-                        f"Multi-bit register '{name}' (bit[{size}]) cannot be used as a "
-                        f"single-bit condition. Use an indexed reference like '{name}[0]'."
-                    )
-            index = 0
-        else:
-            raise TypeError(f"Unsupported condition operand type: {type(node)}")
-
-        return self._clbit_offset[name] + index
-
-
-def native_gate_connectivity(properties: DeviceCapabilities) -> list[list[int]] | None:
-    """Returns the connectivity natively supported by a Braket device from its properties
-
-    Args:
-        properties (DeviceCapabilities): The device properties of the Braket device.
-
-    Returns:
-        list[list[int]] | None: A list of connected qubit pairs or ``None``
-        if the device is fully connected.
-    """
-    device_connectivity = properties.paradigm.connectivity
-    return (
-        [
-            [int(x), int(y)]
-            for x, neighborhood in device_connectivity.connectivityGraph.items()
-            for y in neighborhood
-        ]
-        if not device_connectivity.fullyConnected
-        else None
-    )
-
-
-def native_gate_set(properties: DeviceCapabilities) -> set[str]:
-    """Returns the gate set natively supported by a Braket device from its properties
-
-    Args:
-        properties (DeviceCapabilities): The device properties of the Braket device.
-
-    Returns:
-        set[str]: The names of Qiskit gates natively supported by the Braket device.
-    """
-    native_list = properties.paradigm.nativeGateSet
-    return {
-        _BRAKET_TO_QISKIT_NAMES[op.lower()]
-        for op in native_list
-        if op.lower() in _BRAKET_TO_QISKIT_NAMES
-    }
-
-
-def native_angle_restrictions(
-    properties: DeviceCapabilities,
-) -> dict[str, dict[int, set[float] | tuple[float, float]]]:
-    """Returns angle restrictions for gates natively supported by a Braket device.
-
-    The returned mapping specifies, for each gate name, constraints on the
-    gate parameters indexed by their position. The constraint can either be a
-    set of allowed angles or a tuple representing an inclusive ``(min, max)``
-    range. Angle units are in radians.
-
-    Args:
-        properties (DeviceCapabilities): The device properties of the Braket device.
-
-    Returns:
-        dict[str, dict[int, set[float] | tuple[float, float]]]: Mapping
-        of gate names to parameter index restrictions.
-    """
-
-    if isinstance(properties, (RigettiDeviceCapabilities, RigettiDeviceCapabilitiesV2)):
-        return {"rx": {0: {pi, -pi, pi / 2, -pi / 2}}}
-    if isinstance(properties, IonqDeviceCapabilities):
-        return {"ms": {2: (0.0, 0.25)}}
-    return {}
-
-
-def gateset_from_properties(properties: OpenQASMDeviceActionProperties) -> set[str]:
-    """Returns the gateset supported by a Braket device with the given properties
-
-    Args:
-        properties (OpenQASMDeviceActionProperties): The action properties of the Braket device.
-
-    Returns:
-        set[str]: The names of the gates supported by the device
-    """
-    gateset = {
-        _BRAKET_TO_QISKIT_NAMES[op.lower()]
-        for op in properties.supportedOperations
-        if op.lower() in _BRAKET_TO_QISKIT_NAMES
-    }
-    if "u" in gateset:
-        gateset.update(_ADDITIONAL_U_GATES)
-    max_control = 0
-    for modifier in properties.supportedModifiers:
-        if isinstance(modifier, Control):
-            max_control = modifier.max_qubits
-            break
-    return gateset.union(_get_controlled_gateset(gateset, max_control))
-
-
-def _get_controlled_gateset(base_gateset: set[str], max_qubits: int | None = None) -> set[str]:
-    """Returns the Qiskit gates expressible as controlled versions of existing Braket gates
-
-    This set can be filtered by the maximum number of control qubits.
-
-    Args:
-        base_gateset (set[str]): The base (without control modifiers) gates supported
-        max_qubits (int | None): The maximum number of control qubits that can be used to express
-            the Qiskit gate as a controlled Braket gate. If ``None``, then there is no limit to the
-            number of control qubits. Default: ``None``.
-
-    Returns:
-        set[str]: The names of the controlled gates.
-    """
-    max_control = max_qubits if max_qubits is not None else inf
-    return {
-        controlled_gate
-        for control_count, gate_map in _CONTROLLED_GATES_BY_QUBIT_COUNT.items()
-        for controlled_gate, base_gate in gate_map.items()
-        if control_count <= max_control and base_gate in base_gateset
-    }
-
-
-# TODO: move target construction to a dedicated file; AwsDevice target construction is getting big
-def local_simulator_to_target(simulator: LocalSimulator) -> Target:
-    """Converts properties of a Braket LocalSimulator into a Qiskit Target object.
-
-    Args:
-        simulator (LocalSimulator): Amazon Braket ``LocalSimulator``
-
-    Returns:
-        Target: Target for Qiskit backend
-    """
-    return _simulator_target(
-        simulator, f"Target for Amazon Braket local simulator: {simulator.name}"
-    )
-
-
-def aws_device_to_target(device: AwsDevice) -> Target:
-    """Converts properties of Braket AwsDevice into a Qiskit Target object.
-
-    Args:
-        device (AwsDevice): Amazon Braket ``AwsDevice``
-
-    Returns:
-        Target: Target for Qiskit backend
-    """
-    match device.type:
-        case AwsDeviceType.QPU:
-            return _qpu_target(device, f"Target for Amazon Braket QPU: {device.name}")
-        case AwsDeviceType.SIMULATOR:
-            return _simulator_target(device, f"Target for Amazon Braket simulator: {device.name}")
-    raise QiskitBraketException(
-        "Cannot convert to target. "
-        f"{device.properties.__class__} device capabilities are not supported."
-    )
-
-
-def _simulator_target(device: Device, description: str) -> Target:
-    properties: GateModelSimulatorDeviceCapabilities = device.properties
-    target = Target(description=description, num_qubits=properties.paradigm.qubitCount)
-    action = properties.action.get(DeviceActionType.OPENQASM) or properties.action.get(
-        DeviceActionType.JAQCD
-    )
-    for operation in action.supportedOperations:
-        instruction = _BRAKET_GATE_NAME_TO_QISKIT_GATE.get(operation.lower())
-        if instruction:
-            target.add_instruction(instruction, name=_BRAKET_TO_QISKIT_NAMES[operation.lower()])
-    if isinstance(action, OpenQASMDeviceActionProperties):
-        max_control = 0
-        for modifier in action.supportedModifiers:
-            if isinstance(modifier, Control):
-                max_control = modifier.max_qubits
-                break
-        for gate in _get_controlled_gateset(target.keys(), max_control):
-            if gate in _STANDARD_GATE_NAME_MAPPING:
-                target.add_instruction(_STANDARD_GATE_NAME_MAPPING[gate])
-    target.add_instruction(Measure())
-    target.add_instruction(Barrier(1))
-    return target
-
-
-def _qpu_target(device: AwsDevice, description: str) -> Target:
-    properties: DeviceCapabilities = device.properties
-    topology = device.topology_graph
-    standardized = properties.standardized
-    indices = {q: i for i, q in enumerate(sorted(topology.nodes))}
-
-    qubit_properties = []
-    default_instruction_props = InstructionProperties(error=0)
-    instruction_props_measurement: dict[tuple[int, ...], InstructionProperties] = {}
-    instruction_props_1q: dict[tuple[int, ...], InstructionProperties] = {}
-    instruction_props_2q: dict[str, dict[tuple[int, ...], InstructionProperties]] = {}
-    # TODO: Support V3 standardized properties
-    if isinstance(standardized, (StandardizedPropertiesV1, StandardizedPropertiesV2)):
-        props_1q = standardized.oneQubitProperties
-        for q in sorted(int(q) for q in props_1q):
-            if q not in indices:
-                warnings.warn(
-                    f"Qubit {q} found in device properties but not in topology. "
-                    f"Skipping qubit {q} and its associated properties.",
-                    UserWarning,
-                    stacklevel=2,
-                )
-                continue
-            props = props_1q[str(q)]
-            key = (indices[q],)
-            for fidelity in props.oneQubitFidelity:
-                match fidelity.fidelityType.name.lower():
-                    case "readout":
-                        instruction_props_measurement[key] = InstructionProperties(
-                            # Use highest known error rate
-                            error=max(
-                                1 - fidelity.fidelity,
-                                instruction_props_measurement.get(
-                                    key, default_instruction_props
-                                ).error,
-                            )
-                        )
-                    case name if "readout_error" not in name:
-                        instruction_props_1q[key] = InstructionProperties(
-                            error=max(
-                                1 - fidelity.fidelity,
-                                instruction_props_1q.get(key, default_instruction_props).error,
-                            )
-                        )
-            qubit_properties.append(QubitProperties(t1=props.T1.value, t2=props.T2.value))
-        instruction_props_2q.update(
-            _build_instruction_props_2q(standardized, indices, default_instruction_props)
-        )
-
-    default_props_1q: dict[tuple[int, ...], None] = {(i,): None for i in indices.values()}
-    default_props_2q: dict[tuple[int, ...], None] = {
-        (indices[u], indices[v]): None for u, v in topology.edges
-    }
-    if not instruction_props_measurement:
-        instruction_props_measurement.update(default_props_1q)
-    if not instruction_props_1q:
-        instruction_props_1q.update(default_props_1q)
-
-    parameter_restrictions = _get_parameter_restrictions(device, indices)
-    target = _SubstitutedTarget(
-        description=description,
-        num_qubits=len(qubit_properties or indices),
-        qubit_properties=qubit_properties or None,
-    )
-    if parameter_restrictions:
-        _add_instructions_parameter_restrictions(
-            target,
-            parameter_restrictions,
-            instruction_props_1q,
-            instruction_props_2q,
-            default_props_2q,
-        )
-    else:
-        _add_instructions_no_parameter_restrictions(
-            target,
-            properties.paradigm.nativeGateSet,
-            instruction_props_1q,
-            instruction_props_2q,
-            default_props_2q,
-        )
-
-    if "barrier" in properties.paradigm.nativeGateSet:
-        target.add_instruction(Barrier(1))
-
-    # Add measurement if not already added
-    if "measure" not in target:
-        target.add_instruction(Measure(), instruction_props_measurement)
-    return target
-
-
-def _build_instruction_props_2q(
-    standardized: StandardizedPropertiesV1 | StandardizedPropertiesV2,
-    indices: Mapping[int, int],
-    default_properties: InstructionProperties,
-) -> dict[str, dict[tuple[int, ...], InstructionProperties]]:
-    instruction_props_2q: dict[str, dict[tuple[int, ...], InstructionProperties]] = defaultdict(
-        dict
-    )
-    for k, props in standardized.twoQubitProperties.items():
-        qubits = [int(q) for q in k.split("-")]
-        # Check if all qubits in the edge exist in topology
-        if not all(q in indices for q in qubits):
-            missing_qubits = [q for q in qubits if q not in indices]
-            warnings.warn(
-                f"Edge {k} contains qubits {missing_qubits} not found in topology. "
-                f"Skipping edge {k} and its associated properties.",
-                UserWarning,
-                stacklevel=2,
-            )
-            continue
-
-        for fidelity in props.twoQubitGateFidelity:
-            if gate_name := _BRAKET_TO_QISKIT_NAMES.get(fidelity.gateName.lower()):
-                edge = tuple(indices[q] for q in qubits)
-                instruction_props_2q[gate_name][edge] = InstructionProperties(
-                    error=max(
-                        1 - fidelity.fidelity,
-                        instruction_props_2q[gate_name].get(edge, default_properties).error,
-                    )
-                )
-    # Standardized 2q gate props assume bidirectionality
-    for edge_props in instruction_props_2q.values():
-        edge_props.update({
-            tuple(reversed(edge)): instruction_props
-            for edge, instruction_props in edge_props.items()
-        })
-    return instruction_props_2q
-
-
-def _get_parameter_restrictions(
-    device: AwsDevice, qubit_indices: Mapping[int, int]
-) -> _ParameterRestrictions:
-    cal = device.gate_calibrations
-    parameter_restrictions: _ParameterRestrictions = defaultdict(lambda: defaultdict(set))
-    for gate, target in cal.pulse_sequences if cal else {}:
-        gate_name = gate.name.lower()
-        qubits = tuple(qubit_indices[q] for q in target)
-        if isinstance(gate, Parameterizable):
-            param_key = tuple(
-                param.name if isinstance(param, FreeParameter) else param
-                for param in gate.parameters
-            )
-            parameter_restrictions[gate_name][tuple(param_key)].add(qubits)
-        else:
-            parameter_restrictions[gate_name][()].add(qubits)
-    return parameter_restrictions
-
-
-def _add_instructions_parameter_restrictions(
-    target: _SubstitutedTarget,
-    parameter_restrictions: Mapping[str, Mapping[_ParamKey, _QubitSet]],
-    instruction_props_1q: Mapping[tuple[int, ...], InstructionProperties],
-    instruction_props_2q: Mapping[str, Mapping[tuple[int, ...], InstructionProperties]],
-    default_props_2q: Mapping[tuple[int, ...], InstructionProperties | None],
-) -> None:
-    for braket_name, restrictions in parameter_restrictions.items():
-        if instruction := _BRAKET_GATE_NAME_TO_QISKIT_GATE.get(braket_name):
-            gate_name = instruction.name
-            match num_qubits := instruction.num_qubits:
-                case 1:
-                    _add_single_instruction_parameter_restriction(
-                        target,
-                        instruction,
-                        braket_name,
-                        restrictions,
-                        instruction_props_1q,
-                    )
-                case 2:
-                    _add_single_instruction_parameter_restriction(
-                        target,
-                        instruction,
-                        braket_name,
-                        restrictions,
-                        instruction_props_2q.get(gate_name, default_props_2q),
-                    )
-                case _:
-                    warnings.warn(
-                        f"Instruction {gate_name} has {num_qubits} qubits "
-                        "and cannot be added to target",
-                        stacklevel=2,
-                    )
-
-
-def _add_single_instruction_parameter_restriction(
-    target: _SubstitutedTarget,
-    instruction: QiskitInstruction,
-    braket_name: str,
-    restrictions: Mapping[_ParamKey, _QubitSet],
-    gate_properties: Mapping[tuple[int, ...], InstructionProperties],
-) -> None:
-    for restriction, qubits in restrictions.items():
-        props = {q: props for q, props in gate_properties.items() if q in qubits}
-        instruction_copy = instruction.copy()
-        if restriction:
-            instruction_copy.params = [
-                Parameter(param) if isinstance(param, str) else param for param in restriction
-            ]
-        if substitute := _TRANSPILER_GATE_SUBSTITUTES.get((braket_name, restriction)):
-            substitute_name = substitute.name
-            if instruction_target := target.get(substitute_name):
-                # Nothing to do with Hamiltonian mechanics; variable names are coincidental :)
-                for q, p in props.items():
-                    if not (current := instruction_target.get(q)) or current.error > p.error:
-                        instruction_target[q] = p
-                        target._gate_substitutes[substitute_name][q] = instruction_copy
-            else:
-                target.add_instruction(substitute, props)
-                target._gate_substitutes[substitute_name] = dict.fromkeys(props, instruction_copy)
-        else:
-            target.add_instruction(instruction_copy, props)
-
-
-def _add_instructions_no_parameter_restrictions(
-    target: _SubstitutedTarget,
-    native_gateset: set[str],
-    instruction_props_1q: Mapping[tuple[int, ...], InstructionProperties],
-    instruction_props_2q: Mapping[str, Mapping[tuple[int, ...], InstructionProperties]],
-    default_props_2q: Mapping[tuple[int, ...], InstructionProperties | None],
-) -> None:
-    for operation in native_gateset:
-        if instruction := _BRAKET_GATE_NAME_TO_QISKIT_GATE.get(operation.lower()):
-            gate_name = instruction.name
-            match num_qubits := instruction.num_qubits:
-                case 1:
-                    target.add_instruction(instruction, instruction_props_1q)
-                case 2:
-                    target.add_instruction(
-                        instruction, instruction_props_2q.get(gate_name, default_props_2q)
-                    )
-                case _:
-                    warnings.warn(
-                        f"Instruction {gate_name} has {num_qubits} qubits "
-                        "and cannot be added to target",
-                        stacklevel=2,
-                    )
 
 
 def _extract_verbatim_boxes(
@@ -1462,7 +320,7 @@ def _compile(
                 routing_method=effective_routing_method,
                 seed_transpiler=seed_transpiler,
             )
-    if isinstance(target, _SubstitutedTarget):
+    if isinstance(target, SubstitutedTarget):
         circuits = target._substitute(circuits)
 
     if has_verbatim_boxes:
@@ -1996,12 +854,12 @@ def to_qiskit(
     """
     if isinstance(circuit, Program):
         return (
-            Interpreter(_QiskitProgramContext(verbatim_box_name))
+            Interpreter(QiskitProgramContext(verbatim_box_name))
             .run(circuit.source, inputs=circuit.inputs)
             .circuit
         )
     if isinstance(circuit, str):
-        return Interpreter(_QiskitProgramContext(verbatim_box_name)).run(circuit).circuit
+        return Interpreter(QiskitProgramContext(verbatim_box_name)).run(circuit).circuit
     if not isinstance(circuit, Circuit):
         raise TypeError(f"Expected a Circuit, got {type(circuit)} instead.")
 
@@ -2063,25 +921,6 @@ def _create_qiskit_kraus(gate_params: list[np.ndarray]) -> Instruction:
     return qiskit_qi.Kraus(gate_params)
 
 
-def _sympy_to_qiskit(
-    expr: Expr, param_map: dict[str, Parameter]
-) -> ParameterExpression | Parameter:
-    """convert a sympy expression to qiskit Parameters recursively"""
-    match expr:
-        case Symbol(name=name):
-            if name not in param_map:
-                param_map[name] = Parameter(name)
-            return param_map[name]
-        case Add(args=args):
-            return sum(_sympy_to_qiskit(arg, param_map) for arg in args)
-        case Mul(args=args):
-            return prod(_sympy_to_qiskit(arg, param_map) for arg in args)
-        case Pow(base=base, exp=exp):
-            return _sympy_to_qiskit(base, param_map) ** int(exp)
-        case obj if getattr(obj, "is_real", False):
-            return float(obj)
-    raise TypeError(f"unrecognized parameter type in conversion: {type(expr)}")
-
 
 def _reverse_endianness(matrix: np.ndarray) -> np.ndarray:
     n_q = int(np.log2(matrix.shape[0]))
@@ -2110,7 +949,7 @@ def _create_qiskit_gate(
         param = next(iter(param_expression.parameters)).sympify()
         coeff = float(param_expression.sympify().subs(param, 1))
         new_gate_params.append(
-            _sympy_to_qiskit(coeff * value.expression, param_map)
+            sympy_to_qiskit(coeff * value.expression, param_map)
             if isinstance(value, FreeParameterExpression)
             else coeff * value
         )
